Refactor `CRUD::delete` method to support advanced DELETE queries (#25)
The current implementation relies on `wpdb::delete`, which supports only simple `where x = y` conditions. It's not only incomplete but also dangerous because `Where::$comparisonOperator` and `Where::$logicalOperator` are just ignored, so the result is unpredictable. The following change provides a way to use existing QueryBuilder functionality to generate DELETE SQL. The documentation describes some restrictions.